+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **Auto-enchant rune**
 - **Automatically level up and drop titan hero gift**
 - **Task retries are now persisted**
+- **Retry arena if estimated win probability is too low**
 - `heroes.js` interface
 - Vendored game resources
 - New scheduler